feat: finish Event Manager MVP, refactore some common logic.
implement an ability to edit the existing event.
implement an ability to update the events list with button.
implement an ability to set auto update time (in sec) for events list.
implement an ability to reset all changes to last saved ones.
implement a logic to save/load Events Manager (works with .txt files)
implement the same logic for the context pop-up menu in the Event Manager as in the Contacts book.
implement a simple instructions dialog for the Events Manager usage.
implement a simple alarm for events (messagebox with a system sound notification).
move the common logic for work with files from the Contact book to Utils.
move the common logic for registering/unregistering hotkeys from the Contact book to HotkeysHandler.
remove some old scripts.